Add javadoc for autonomous classes and organize them into packages
