+++ src/xbot/common/autonomous/ScriptedCommandFactory.java
@@ -1,5 +1,10 @@
 package xbot.common.autonomous;
 
+/**
+ * An interface which describes a factory which translates command
+ * type names into {@link ScriptedCommandProvider}s.
+ *
+ */
 public interface ScriptedCommandFactory {
     public ScriptedCommandProvider getProviderForName(String commandTypeName);
 }
