+++ src/xbot/common/autonomous/ScriptedCommandProvider.java
@@ -2,6 +2,12 @@ package xbot.common.autonomous;
 
 import edu.wpi.first.wpilibj.command.Command;
 
+
+/**
+ * An interface which describes a provider which translates configuration
+ * options into a Command instance.
+ *
+ */
 public interface ScriptedCommandProvider {
     public Command get();
 }
